[OT-183] [REFACTOR]: 콘텐츠 관련 hooks 하나의 파일로 통합
